+++ jobbergate-cli/CHANGELOG.rst
@@ -6,7 +6,7 @@ This file keeps track of all notable changes to jobbergate-cli
 
 Unreleased
 ----------
-* Move SBATCH params to Job Script submission
+* Added support for `execution_parameters` in job submission at the CLI
 
 3.3.4 -- 2022-12-05
 -------------------